Move building cluster from match table to "cluster" code
